% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichment_test}
\alias{enrichment_test}
\title{enrichment test}
\usage{
enrichment_test(X, design, group, coef, test, p.cutoff = 1)
}
\arguments{
\item{X}{The concentration/expression matrix. Rows are variables and columns
are observations.}

\item{design}{The design matrix often generated from
\code{\link{model.matrix}}. Number of rows must equal to the number of
columns of X.}

\item{group}{character. The categories of variables. Length must equal to the
number of rows of X.}

\item{coef}{character. The main effect of interest in the design matrix. Must
be one of the column names of the design matrix.}

\item{test}{character. Either hyper or ks.}

\item{p.cutoff}{numeric. The p-value cutoff to be used in hyper test. Default
is 1.}
}
\value{
A object of class either "enrichment.hyper" or "enrichment.ks" is
returned by this function. The returned object has the elements as described
below.

\describe{
\item{pvalue}{The pvalue of the enrichment test.}
\item{lm.fit}{The linear model result. The pvalues are adjusted for ks.test}
\item{group}{The categrories of variables}
\item{hyper.matrix}{enrichment.hyper only. The hypergeometric test parameters
for each variable.}
\item{p.cutoff}{enrichment.hyper only. The pvalue cutoff to define
whether a variable was affected or not.}
\item{d}{enrichment.ks only. The value of ks.test statistics.
See \code{\link{ks.test}}}.
}
\item{adj.p.val}{enrichment.ks only. The adjusted pvalues for each variable}
}
\description{
Perform enrichment test on a concentration/expression matrix
of a hight through-put experiment study.

A linear model was first fitted for each variable against the provided design
matrix. The statistics were then used to perform either the hypergeometric
test (Fisher's exact test) or Kolmogorov-Smirnov test to see whether a
particular category of variables are enriched.

The function \code{\link{phyper}} is used for hypergeometric test for each
category in the parameter \strong{group}. For any given category \strong{g}
in \strong{group}, parameters for \code{\link{phyper}} are defined as
following. m is the total number of variables belong to category \strong{g}.
n is the total number of variables not belong to category \strong{g}. k is
the total number of variables that were affected in the study. x is the
number of variables belong to category \strong{g} and also affected.
Variables are defined as affected if the estimated coefficient is higher or
lower according to the parameter \strong{alternative}, with a p-value
smaller than the cutoff (p.cutoff).

The function \code{\link{ks.test}} is used for Kolmogorov-Smirnov test for
each category in \strong{group}. The pvalues were first adjusted to one-tail
using \code{\link{pt}} according to the parameter \strong{alternative}. The
Kolmogorov-Smirnov test is then performed between the adjusted p-values of
variables that belong to the category \strong{g} and a sequence of simulated
values from 0 to 1.
}
\examples{
data(glyco)
X = glyco$conc_table
design = model.matrix(~flipgroup, data = as(glyco$sample_table, "data.frame"))
group = glyco$feature_data$Protein
enrichment_test(X, design, group, "flipgroupcLNS", "hyper")
enrichment_test(X, design, group, "flipgroupcLNS", "ks")

}
\seealso{
\code{\link{ks.test}} \code{\link{phyper}}
}
\author{
Chenghao Zhu
}
