% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco.R
\docType{data}
\name{glyco}
\alias{glyco}
\title{glycoproteome dataset}
\format{An object of class \code{GlycomicsSet} with 163 rows and 80 columns.}
\usage{
glyco
}
\description{
This is a dataset of a glycoproteome study that looked at 80
human subjects under two different tweatment. The data was an object of
the GlycomicsSet class defined in the Metabase package. And the Metabase
package is required in order to use this dataset. The package can be
installed from the github (https://github.com/zhuchcn/Metabase).
}
\keyword{datasets}
