% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{plot.enrichment.hyper}
\alias{plot.enrichment.hyper}
\title{enrichment barplot}
\usage{
\method{plot}{enrichment.hyper}(x, each = 3, ...)
}
\arguments{
\item{x}{enrichment.hyper. Returned by \code{\link{enrichement_test}}
when test was set to hyper.}

\item{each}{integer. The top most enriched levels on each side.}
}
\value{
a ggplot object
}
\description{
enrichment barplot

This function makes bar plot for top most positively and negatively
enriched levels.
}
\seealso{
\code{\link{enrichment_test}}
}
\author{
Chenghao Zhu
}
